Create Author Create Book Create Membership List all books Find info about specific book

+++ LibraryApplicationProject/Data/Membership.cs
@@ -5,7 +5,7 @@ namespace LibraryApplicationProject.Data;
 public class Membership
 {
     public int Id { get; set; }
-    public int CardNumber { get; set; }
+    public long CardNumber { get; set; }
     public DateTime RegistryDate { get; set; }
     public DateTime? ExpirationDate { get; set; }
     public Person? Person { get; set; }
